refactor: update missed names
