+++ src/test/java/com/festp/styledmessage/attributes/FormattingTests.java
@@ -1,4 +1,4 @@
-package com.festp.styledmessage.components;
+package com.festp.styledmessage.attributes;
 
 import org.bukkit.ChatColor;
 import org.junit.jupiter.api.Assertions;
